frontend: 3d visualizer
