+++ frontend/react/src/pages/visualizacao3D/dataFetcher.ts
@@ -1,17 +1,5 @@
 import { Empenho3DItem } from "./types";
 
-// export async function fetchEmpenhos3D(): Promise<Empenho3DItem[]> {
-//   try {
-//     const response = await fetch("http://localhost:8000/api/empenhos-3d");
-//     if (!response.ok) throw new Error("Erro ao buscar dados 3D");
-//     return await response.json();
-//   } catch (error) {
-//     console.error(error);
-//     return [];
-//   }
-// }
-
-
 export async function fetchAllEmpenhos3D(empenhoId : string): Promise<Empenho3DItem[]> {
   try {
     const payload = { empenhoId: empenhoId };
@@ -27,3 +15,29 @@ export async function fetchAllEmpenhos3D(empenhoId : string): Promise<Empenho3DI
     return [];
   }
 }
+
+export const fetchAutoComplete = async (query: string, type: number) => {
+    if (!query.trim()) {
+      return [];
+    }
+
+    try {
+      const payload = { consulta: query, tipo: type };
+      const response = await fetch("http://localhost:8000/api/auto-filling", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(payload),
+      });
+
+      const data = await response.json();  
+      return data;
+
+    } catch (err) {
+      console.error("Erro ao buscar sugestões:", err);
+      return err;
+    } 
+    // finally {
+    //   setLoading(false);
+    // }
+  };
+
